Handle unhandled conditions from user-defined primitives.
When there are such conditions, the error is stored in the
'user-primitive-error' field of the virtual machine and the virtual
machine goes to sleep as if :nap-and-retry-on-wake-up had been
returned by the primitive as a second value. This gives the user the
chance to fix the user-defined primitive or the context.