More preprocessing on instructions when executing them. Changed tests.
I'm trying to organize instructions in 'runs' which will execute
faster (right now because the quotas are not checked during a run and
the ticks counter is not incremented during a run).

Modified tests to match the new way of incrementing/checking quotas.

Modified documentation (it said that we update and check quotas on
every instruction, which is no longer true).