C# implementation: optimizations.
Took 'Cache' away from Instruction so the same array of Instruction
can be used and reused by many VMs (turns out that deserializing
bytecode is expensive).

Pseudo-jitting the bytecode (calculating runs) is now only enabled if
the instruction at the start of the run is executed more than 10
times. This saves pseudo-jitting time for code that doesn't run very
often.