C# implementation: fixed current environment dumping.
If an error occurs in a program made of multiple files, the global
frame of the environment spans multiple files and the attempt to
display its lines fails because the end position may not be found in
the file where the frame starts.

This fix changes the way environments are displayed so this problem
doesn't occur.